tq_api_backend-main/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
dictdiffer==0.9.0
execnet==1.9.0
filelock==3.7.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
Jinja2==3.1.2
jsonpath==0.82
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.21.6
opencv-python==4.5.4.58
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
PyAutoGUI==0.9.53
PyAutoIt==0.6.5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.0.0
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-parallel==0.0.10
pytest-rerunfailures==10.2
pytweening==1.0.4
pywin32==304
requests==2.27.1
six==1.16.0
smmap==5.0.0
tblib==1.7.0
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0

xlrd2==1.3.4
psutil==5.9.1
WMI==1.5.1